auctioning_platform/processes/requirements.txt
freezegun==0.3.12
pytest-freezegun==0.3.0.post1
typing-extensions==3.7.2
injector==0.16.2
SQLAlchemy==1.3.4
